requirements.txt
flask==3.1.0
gunicorn==223.0.0
agno==1.2.10 
python-dotenv==1.1.0
pgvector==0.4.0 
openai==1.70.0
pypdf==5.4.0
requests==2.32.3
urllib3==2.3.0
psycopg2-binary==2.9.10
googlesearch-python==1.3.0
pycountry==24.6.1
sqlalchemy==2.0.40